pyproject.toml
[tool.poetry]
name = "async-architecture-ates"
version = "0.1.0"
description = ""
authors = ["Sergey"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
jinja2 = "^3.1.3"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
pydantic-settings = "^2.2.1"
pydantic = "^2.6.2"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
typer = {extras = ["all"], version = "^0.9.0"}
authlib = "^1.3.0"
httpx = "^0.27.0"
aiokafka = "^0.10.0"
faststream = {extras = ["kafka"], version = "^0.4.5"}


[tool.poetry.group.dev.dependencies]
ruff = "^0.3.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
# Enable Pyflakes (`F`) and a subset of the pycodestyle (`E`)  codes by default.
select = ["E4", "E7", "E9", "F"]
ignore = []

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "single"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"


[tool.isort]
profile = "hug"
line_length = 100
force_single_line = true
